Expose loggers logic
