+++ python/arcticdb/log.py
@@ -8,54 +8,67 @@ As of the Change Date specified in that file, in accordance with the Business So
 
 import traceback
 
-from arcticdb_ext.log import configure
-from arcticdb_ext.log import log as _log, LogLevel as _Lvl, LoggerId as _LoggerId, is_active as _is_active
+
+def build_loggers(ext_log):
+    """Return a logger registry bound to the given extension's log submodule.
+
+    ``ext_log`` must expose ``log``, ``LogLevel``, ``LoggerId`` and ``is_active``
+    (see ``register_log`` in ``python_bindings_common.cpp``).
+    """
+    _log = ext_log.log
+    _Lvl = ext_log.LogLevel
+    _is_active = ext_log.is_active
+    _LoggerId = ext_log.LoggerId
+
+    class _Logger:
+        def __init__(self, id):
+            self._id = id
+
+        def log(self, lvl, msg, *args, **kwargs):
+            if not self.is_active(lvl):
+                return
+            _log(self._id, lvl, msg.format(*args, **kwargs))
+
+        def debug(self, msg, *args, **kwargs):
+            self.log(_Lvl.DEBUG, msg, *args, **kwargs)
+
+        def info(self, msg, *args, **kwargs):
+            self.log(_Lvl.INFO, msg, *args, **kwargs)
+
+        def warning(self, msg, *args, **kwargs):
+            self.log(_Lvl.WARN, msg, *args, **kwargs)
+
+        def is_active(self, lvl):
+            return _is_active(self._id, lvl)
+
+        warn = warning
+
+        def error(self, msg, *args, **kwargs):
+            self.log(_Lvl.ERROR, msg, *args, **kwargs)
+
+        def exception(self, msg, *args, **kwargs):
+            exc = traceback.format_exc()
+            _log(self._id, _Lvl.ERROR, msg.format(*args, **kwargs) + "\n" + exc)
+
+    return {
+        "codec": _Logger(_LoggerId.CODEC),
+        "inmem": _Logger(_LoggerId.IN_MEM),
+        "root": _Logger(_LoggerId.ROOT),
+        "storage": _Logger(_LoggerId.STORAGE),
+        "version": _Logger(_LoggerId.VERSION),
+        "memory": _Logger(_LoggerId.MEMORY),
+        "timings": _Logger(_LoggerId.TIMINGS),
+        "lock": _Logger(_LoggerId.LOCK),
+        "schedule": _Logger(_LoggerId.SCHEDULE),
+        "symbol": _Logger(_LoggerId.SYMBOL),
+        "snapshot": _Logger(_LoggerId.SNAPSHOT),
+    }
 
 
-class _Logger(object):
-    def __init__(self, id):
-        self._id = id
+import arcticdb_ext.log as _ext_log
 
-    def log(self, lvl, msg, *args, **kwargs):
-        if not self.is_active(lvl):
-            return
-        _log(self._id, lvl, msg.format(*args, **kwargs))
+configure = _ext_log.configure
+logger_by_name = build_loggers(_ext_log)
 
-    def debug(self, msg, *args, **kwargs):
-        self.log(_Lvl.DEBUG, msg, *args, **kwargs)
-
-    def info(self, msg, *args, **kwargs):
-        self.log(_Lvl.INFO, msg, *args, **kwargs)
-
-    def warning(self, msg, *args, **kwargs):
-        self.log(_Lvl.WARN, msg, *args, **kwargs)
-
-    def is_active(self, lvl):
-        return _is_active(self._id, lvl)
-
-    warn = warning
-
-    def error(self, msg, *args, **kwargs):
-        self.log(_Lvl.ERROR, msg, *args, **kwargs)
-
-    def exception(self, msg, *args, **kwargs):
-        exc = traceback.format_exc()
-        _log(self._id, _Lvl.ERROR, msg.format(*args, **kwargs) + "\n" + exc)
-
-
-logger_by_name = {
-    "codec": _Logger(_LoggerId.CODEC),
-    "inmem": _Logger(_LoggerId.IN_MEM),
-    "root": _Logger(_LoggerId.ROOT),
-    "storage": _Logger(_LoggerId.STORAGE),
-    "version": _Logger(_LoggerId.VERSION),
-    "memory": _Logger(_LoggerId.MEMORY),
-    "timings": _Logger(_LoggerId.TIMINGS),
-    "lock": _Logger(_LoggerId.LOCK),
-    "schedule": _Logger(_LoggerId.SCHEDULE),
-    "symbol": _Logger(_LoggerId.SYMBOL),
-    "snapshot": _Logger(_LoggerId.SNAPSHOT),
-}
-
-for key, value in logger_by_name.items():
-    globals()[key] = value
+for _key, _value in logger_by_name.items():
+    globals()[_key] = _value
